+++ NEWS.md
@@ -7,6 +7,7 @@
 * Smarter string ID handling for ped2id
 * Fixed how different-sized matrices are handled by `com2links()`
 * Added alignPhenToMatrix function to align phenotypic data to the order of the relatedness matrix
+* Added `simulatePedigrees()` function to easily simulate multiple families at once and return them as a single combined data frame
 
 # BGmisc 1.7.0.0
 * Fixed bug in parList